Moved stats from controller to new PlayerInformation class
