+++ Assets/02.Scripts/FSM/State/StateController/PlayerStateController.cs
@@ -6,11 +6,6 @@ using IEGame.FiniteStateMachine;
 [ExecuteInEditMode]
 public class PlayerStateController : MonoBehaviour, IStateController
 {
-    [field: SerializeField] public int Level { get; set; }
-    [field: SerializeField] public int Score { get; set; }
-    [field: SerializeField] public int Money { get; set; }
-    [field: SerializeField] public float Health { get; set; }
-
     #region state
     [field: SerializeField]
     public State CurrentState { get; set; }
@@ -279,30 +274,6 @@ public class PlayerStateController : MonoBehaviour, IStateController
             CurrentState = nextState;
             this.OnExitState();
         }
-    }
-
-    public void SavePlayer()
-    {
-        Debug.Log("Button save going through");
-        SaveSystem.SavePlayer(this);
-    }
-
-    public void LoadPlayer()
-    {
-        PlayerStats data = SaveSystem.LoadPlayer();
-
-        Vector3 position;
-        position.x = data.position[0];
-        position.y = data.position[1];
-        position.z = data.position[2];
-
-        transform.position = position;
-        this.Health = data.Health;
-        this.Score = data.Score;
-        this.Level = data.Level;
-        this.Score = data.Score;
-
-        Debug.Log($"Player was LOADED. Health:{this.Health} Level:{this.Level} Money:{this.Money} Score:{this.Score} Position: x={this.transform.position[0]} y={this.transform.position[1]} y={this.transform.position[2]}");
     }
 }
