--- Assets/02.Scripts/FSM/Stats/SaveSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4868fc275fbb7ad428b162c49da0fa37
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
