+++ Assets/02.Scripts/SaveLoad.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 44b9b49427afd9a4aba98d638a048c29
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
