+++ Assets/02.Scripts/SaveLoad/PlayerInformation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b046507b0c37dca4d84b55b71ec09810
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
