+++ Assets/Test/US8Tests_SaveLoad/SaveLoadTest.cs
@@ -6,25 +6,28 @@ namespace US8Tests_SaveLoad
 {
     public class SaveLoadTest
     {
-        private PlayerStateController playerStateController;
+        private PlayerInformation playerInformation;
 
         [SetUp]
         public void Setup()
         {
+            GameObject player = new GameObject();
+            playerInformation = player.AddComponent<PlayerInformation>();
+
             // first initialise the player's stats
             Vector3 position1 = new Vector3(1f, -7f, 2.55f);
-            playerStateController.Transform.position = position1;
-            playerStateController.Health = 97.2f;
-            playerStateController.Level = 10;
-            playerStateController.Score = 3501;
-            playerStateController.Money = 999999;
+            playerInformation.transform.position = position1;
+            playerInformation.Health = 97.2f;
+            playerInformation.Level = 10;
+            playerInformation.Score = 3501;
+            playerInformation.Money = 999999;
         }
 
         // Test if the save file exists in the location 
         [Test]
         public void TestSavePlayer_SaveFileExistsAfterSaving()
         {
-            SaveSystem.SavePlayer(playerStateController);
+            SaveSystem.SavePlayer(playerInformation);
 
             // Path.GetFileName() will return a null if not found
             Assert.IsNotNull(Path.GetFileName(Application.persistentDataPath + "/playerSaveFile"));
@@ -37,13 +40,13 @@ namespace US8Tests_SaveLoad
             SaveSystem.LoadPlayer();
 
             //check for x position
-            Assert.AreEqual(1f, playerStateController.Transform.position[0]);
+            Assert.AreEqual(1f, playerInformation.transform.position[0]);
 
             //check for y position
-            Assert.AreEqual(-7f, playerStateController.Transform.position[1]);
+            Assert.AreEqual(-7f, playerInformation.transform.position[1]);
 
             //check for z position
-            Assert.AreEqual(2.55f, playerStateController.Transform.position[2]);
+            Assert.AreEqual(2.55f, playerInformation.transform.position[2]);
         }
 
         // Test if the loaded character has the correct health points
@@ -53,7 +56,7 @@ namespace US8Tests_SaveLoad
             SaveSystem.LoadPlayer();
 
             //check for health points
-            Assert.AreEqual(97.2f, playerStateController.Health);
+            Assert.AreEqual(97.2f, playerInformation.Health);
 
         }
 
@@ -64,7 +67,7 @@ namespace US8Tests_SaveLoad
             SaveSystem.LoadPlayer();
 
             //check for health points
-            Assert.AreEqual(10, playerStateController.Level);
+            Assert.AreEqual(10, playerInformation.Level);
 
         }
 
@@ -75,7 +78,7 @@ namespace US8Tests_SaveLoad
             SaveSystem.LoadPlayer();
 
             //check for health points
-            Assert.AreEqual(97.2f, playerStateController.Score);
+            Assert.AreEqual(3501, playerInformation.Score);
 
         }
 
@@ -86,7 +89,7 @@ namespace US8Tests_SaveLoad
             SaveSystem.LoadPlayer();
 
             //check for health points
-            Assert.AreEqual(999999, playerStateController.Money);
+            Assert.AreEqual(999999, playerInformation.Money);
 
         }
     }
